fix contrastive loss
