+++ fastreid/evaluation/__init__.py
@@ -1,6 +1,8 @@
 from .evaluator import DatasetEvaluator, inference_context, inference_on_dataset
 from .reid_evaluation import ReidEvaluator
 from .clas_evaluator import ClasEvaluator
+from .pair_distance_evaluator import PairDistanceEvaluator
+from .pair_score_evaluator import PairScoreEvaluator
 from .testing import print_csv_format, verify_results
 
 __all__ = [k for k in globals().keys() if not k.startswith("_")]
