+++ projects/FastShoe/fastshoe/trainer.py
@@ -13,7 +13,7 @@ from fastreid.data.datasets import DATASET_REGISTRY
 from fastreid.utils import comm
 from fastreid.data.transforms import build_transforms
 from fastreid.data.build import build_reid_train_loader, build_reid_test_loader
-from fastreid.evaluation.pair_score_evaluator import PairScoreEvaluator
+from fastreid.evaluation import PairScoreEvaluator, PairDistanceEvaluator
 from projects.FastShoe.fastshoe.data import PairDataset
 
 
@@ -84,4 +84,4 @@ class PairTrainer(DefaultTrainer):
     @classmethod
     def build_evaluator(cls, cfg, dataset_name, output_dir=None):
         data_loader = cls.build_test_loader(cfg, dataset_name)
-        return data_loader, PairScoreEvaluator(cfg, output_dir)
+        return data_loader, PairDistanceEvaluator(cfg, output_dir)
